requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
Pillow>=10.0.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.21.0
einops>=0.7.0
timm>=0.9.0
av>=10.0.0
opencv-python>=4.8.0
decord>=0.6.0